# swagger.json
{
  "basePath": "/",
  "definitions": {},
  "info": {
    "description": "OpenAPI 3.0 Specification for the REST-API backend of the Amazon EC2 Hosted ML Model",
    "title": "Swagger API - Bentley Assignment",
    "version": "3.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/predict": {
      "get": {
        "description": "Press the `Try it out` button below to test out sample data against the backend REST API.<br/><br/>*Notes:*<br/><br/>**GET:** Used in this demo only; OpenAPI 3.0 standards no longer supports body requests for GET. Returns: `STRING`<br/><br/>**POST:** Used by the front-end service; will pass in parameters via POST. Returns: `JSON`<br/>",
        "parameters": [
          {
            "default": "Building 10",
            "description": "Name of a Building on the Main NIH Campus that is the location of work for an employee.",
            "in": "query",
            "name": "building",
            "required": true,
            "type": "string"
          },
          {
            "default": 20,
            "description": "The total amount of peer-reviewed publications held by an employee.",
            "in": "query",
            "name": "num_publications",
            "required": true,
            "type": "number"
          },
          {
            "default": 30,
            "description": "The total amount of NIH-sanctioned conferences attended by an employee.",
            "in": "query",
            "name": "num_conferences",
            "required": true,
            "type": "number"
          },
          {
            "default": "Fulltime",
            "description": "The classification of an employee as either 'Parttime' or 'Fulltime' (40 hours/week).",
            "in": "query",
            "name": "status",
            "required": true,
            "type": "string"
          },
          {
            "default": "Dr",
            "description": "The official salutation of an employee; may correspond to educational level.",
            "in": "query",
            "name": "title",
            "required": true,
            "type": "string"
          },
          {
            "default": "Infectious Diseases",
            "description": "The specific field of expertise held by an employee.",
            "in": "query",
            "name": "expertise",
            "required": true,
            "type": "string"
          },
          {
            "default": "NHLBI",
            "description": "One of the 27 Institutes of the National Institutes of Health where the employee works.",
            "in": "query",
            "name": "institute",
            "required": true,
            "type": "string"
          },
          {
            "default": 3,
            "description": "The total amount of Postdoctoral Fellows that an employee oversees.",
            "in": "query",
            "name": "num_postdocs",
            "required": true,
            "type": "number"
          },
          {
            "default": 7,
            "description": "The total amount of individuals (contractors and Federal workers) overseen by the employee.",
            "in": "query",
            "name": "num_reports",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "The output values of the request"
          }
        },
        "summary": "Click here to test out sample data.",
        "tags": [
          {
            "name": "Test Prediction"
          }
        ]
      },
      "post": {
        "description": "Press the `Try it out` button below to test out sample data against the backend REST API.<br/><br/>*Notes:*<br/><br/>**GET:** Used in this demo only; OpenAPI 3.0 standards no longer supports body requests for GET. Returns: `STRING`<br/><br/>**POST:** Used by the front-end service; will pass in parameters via POST. Returns: `JSON`<br/>",
        "parameters": [
          {
            "default": "Building 10",
            "description": "Name of a Building on the Main NIH Campus that is the location of work for an employee.",
            "in": "query",
            "name": "building",
            "required": true,
            "type": "string"
          },
          {
            "default": 20,
            "description": "The total amount of peer-reviewed publications held by an employee.",
            "in": "query",
            "name": "num_publications",
            "required": true,
            "type": "number"
          },
          {
            "default": 30,
            "description": "The total amount of NIH-sanctioned conferences attended by an employee.",
            "in": "query",
            "name": "num_conferences",
            "required": true,
            "type": "number"
          },
          {
            "default": "Fulltime",
            "description": "The classification of an employee as either 'Parttime' or 'Fulltime' (40 hours/week).",
            "in": "query",
            "name": "status",
            "required": true,
            "type": "string"
          },
          {
            "default": "Dr",
            "description": "The official salutation of an employee; may correspond to educational level.",
            "in": "query",
            "name": "title",
            "required": true,
            "type": "string"
          },
          {
            "default": "Infectious Diseases",
            "description": "The specific field of expertise held by an employee.",
            "in": "query",
            "name": "expertise",
            "required": true,
            "type": "string"
          },
          {
            "default": "NHLBI",
            "description": "One of the 27 Institutes of the National Institutes of Health where the employee works.",
            "in": "query",
            "name": "institute",
            "required": true,
            "type": "string"
          },
          {
            "default": 3,
            "description": "The total amount of Postdoctoral Fellows that an employee oversees.",
            "in": "query",
            "name": "num_postdocs",
            "required": true,
            "type": "number"
          },
          {
            "default": 7,
            "description": "The total amount of individuals (contractors and Federal workers) overseen by the employee.",
            "in": "query",
            "name": "num_reports",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "The output values of the request"
          }
        },
        "summary": "Click here to test out sample data.",
        "tags": [
          {
            "name": "Test Prediction"
          }
        ]
      }
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0"
}